+++ src/resource/image.ts
@@ -6,4 +6,12 @@ export default {
   empty_img: require("./media/images/empty_img.png"),
   emptyUpload: require("./media/empties/uploadImg.png"),
   reward: require("./media/images/reward.png"),
+  emptyImgReward: require("./media/empties/emptyImgReward.png"),
+  progressBar: require("./media/images/progressBar.png"),
+  uploadImgQuota : require("./media/empties/upload_img_quota.png"),
+  condition : require("./media/images/condition.png"),
+  tableLucky: require("./media/images/tableLucky.png"),
+  announce: require("./media/images/announcement.png"),
+  empty_cover: require("./media/empties/empty-cover.png"),
+  upload_Img_btn : require("./media/empties/upload_Img_btn.png")
 };