+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Version - 0]
+## Feature [v0.34.0] 2023-06-12
+- added quota setting menu
 ## Feature [v0.33.0] 2023-06-12
 - added mission setting menu
 ## Fix [v0.32.8] 2023-06-09